feat: finishing all tests

+++ src/main/java/com/ignite/igniteshop/services/ProductService.java
@@ -9,7 +9,6 @@ import com.ignite.igniteshop.repositories.ProductRepository;
 import com.ignite.igniteshop.services.exceptions.DataBaseException;
 import com.ignite.igniteshop.services.exceptions.ResourceNotFoundException;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -62,10 +61,11 @@ public class ProductService {
     }
 
     public void delete(Long id) {
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Id not found " + id);
+        }
         try {
             productRepository.deleteById(id);
-        } catch (EmptyResultDataAccessException e) {
-            throw new ResourceNotFoundException("Id not found " + id);
         } catch (DataIntegrityViolationException e) {
             throw new DataBaseException("Integrity violation " + id);
         }
